checks

+++ webp2gif.sh
@@ -1,4 +1,20 @@
 #!/bin/bash
-brew install imagemagick
-magick mogrify -format gif ~/Downloads/*.webp
-rm -f ~/Downloads/*.webp
+# Check for dependency
+echo 'Checking for imagemagick...'
+INSTALLED=`brew list imagemagick`
+if [ -n "$INSTALLED" ]; then
+	echo 'Already installed'
+else
+	echo 'Installing...'
+	brew install imagemagick
+fi;
+# Check succcess of last command
+if [[ $?==0 ]]; then
+	echo 'Running mogrify on ~/Downloads/*.webp...'
+	magick mogrify -format gif ~/Downloads/*.webp
+fi;
+# Check succcess of last command
+if [[ $?==0 ]]; then
+	echo 'Removing ~/Downloads/*.webp files...'
+	rm -f ~/Downloads/*.webp
+fi;
